backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
boto3==1.36.4
botocore==1.36.4
pydantic==2.10.5
networkx==3.4.2
faiss-cpu
numpy>=2.0.0
pytesseract==0.3.13
Pillow==11.1.0
PyPDF2==3.0.1
python-docx==1.1.2
aiofiles==24.1.0
